+++ lilya/contrib/openapi/datastructures.py
@@ -1,9 +1,9 @@
 from typing import Annotated, Any
 
 from pydantic import BaseModel, field_validator
-from typing_extensions import Doc
 
 from lilya.enums import MediaType
+from lilya.types import Doc
 
 
 class OpenAPIResponse(BaseModel):